requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
Appium-Python-Client==4.0.0
attrs==23.1.0
axe-selenium-python==2.1.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
h11==0.14.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.3
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==23.2
pip==23.2.1
pluggy==1.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pytest==7.4.3
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-ordering==0.6
requests==2.31.0
selenium==4.16.0
setuptools==69.5.1
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.2.1
wsproto==1.2.0
